+++ MovieWizardAPI.Models/Transaction.cs
@@ -3,7 +3,7 @@
     public class Transactions
     {
         public string? TransactionId {  get; set; }
-        public double? Amount { get; set; }
+        public int? Amount { get; set; }
         public DateTime? CreatedOn { get; set; }
         public int? InvoiceNumber { get; set; }
         public string? CreatedBy { get; set; }
